pyproject.toml
[tool.poetry]
name = "biglib"
version = "1.4.0"
description = "An all encompassing mega-library of stuff"
authors = ["Garry A Offord"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^19.3.0"
fuzzywuzzy = "^0.17.0"
requests = "^2.22.0"
urllib3 = "^1.25.7"
mako = "^1.1.0"
MarkupSafe = "^1.1.1"
python-levenshtein = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
